Cmly: expose Lr1.forbid_default_reduction
Previously, the default_reduction was supposed to be derived from
the transitions and reductions of a state: it has a default reduction if
it had shifts (transition on terminals) and all reductions targetted a
single production.

However, this is incorrect in presence of conflicts resolved using
%nonassoc annotation: this is a shift/reduce conflict that is resolved
in favor of neither shift nor reduce.

This removes one transition and one reduction, making it possible for
somes states to have a default reduction when they should not.