Fix LowIntegerPriorityQueue so that it works in all scenarios.
The previous code would work correctly *only* if the user used an increasing
sequence of priorities, that is, if [remove] at priority [p] can be followed
only by [add] at priorities [p] or greater.

Unfortunately, this was undocumented (and probably unintentional?).

Fortunately, the only client (LRijkstraClassic) respects this constraint,
so there was no bug in Menhir.

The priority queue should now be fixed and work in all scenarios.